Move chapter 4 listings to standard file layout.
