+++ src/main/kotlin/chapter4/either.kt
@@ -4,5 +4,6 @@ package chapter4
 sealed class Either<out E, out A>
 
 data class Left<out E>(val value: E) : Either<E, Nothing>()
+
 data class Right<out A>(val value: A) : Either<Nothing, A>()
 //end::init[]
